In build.gradle set compatibility to 1.8 so jdk 9+ can be used to compile this project with binary compatibility Java 8
Modified UnfairExecutor#doExecute(Runnable) to cast schedulers.length to (long) to ensure 1.8 API is used as a Math.floorMod overload was added for Java 1.9 with (int) second parameter

+++ src/main/java/org/threadly/concurrent/UnfairExecutor.java
@@ -241,7 +241,8 @@ public class UnfairExecutor extends AbstractSubmitterExecutor {
       throw new RejectedExecutionException("Pool is shutdown");
     }
     
-    schedulers[(int)Math.floorMod(stripeGenerator.getStripe(task), schedulers.length)].addTask(task);
+    // Cast schedulers.length to (long) to ensure 1.8 API is used as a Math.floorMod overload was added for Java 1.9 with (int) second parameter
+    schedulers[(int)Math.floorMod(stripeGenerator.getStripe(task), (long)schedulers.length)].addTask(task);
   }
 
   /**
